python_requirements.txt
flask==2.3.3
flask-cors==3.0.10
tensorflow==2.12.0
# opencv-python==4.8.0.76
numpy==1.23.5
pillow==9.5.0
requests==2.31.0
